move platforms to core
