+++ include/fast_io_core_impl/allocation/nt_preliminary_definition.h
